requirements.txt
sounddevice
scipy
numpy
librosa
pydub
torch
torchaudio
torchvision
requests
flask
flask-socketio
matplotlib
gunicorn
ffmpeg-python
mediapipe
pyaudio
pyVBAN